// src/api/generated/model/courseRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Greenfield College System API
 * Complete API Specification covering all mutations by REST.
 * OpenAPI spec version: 1.0.0
 */

export interface CourseRequest {
  school_public_id: string;
  code: string;
  title: string;
  description?: string;
  credits: number;
}