// src/api/generated/model/programRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Greenfield College System API
 * Complete API Specification covering all mutations by REST.
 * OpenAPI spec version: 1.0.0
 */

export interface ProgramRequest {
  school_public_id: string;
  code: string;
  title: string;
  qualification: string;
  duration_months: number;
}