// src/api/generated/model/schoolRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Greenfield College System API
 * Complete API Specification covering all mutations by REST.
 * OpenAPI spec version: 1.0.0
 */

export interface SchoolRequest {
  name: string;
  description?: string;
  /** @nullable */
  dean_user_public_id?: string | null;
}